Remove introductory “Ceylon” slide
That was a lot of information that might not be understandable for
everyone (e. g. imperative vs. functional programming) and that’s not
relevant to my talk. My talk is about the awesome type system,
mentioning that everything else is awesome too belongs at the end, not
at the beginning.